unifying ingest methods
structuring how the changesets will be made
